+++ CHANGELOG.md
@@ -6,7 +6,7 @@ These changes are awaiting release:
 
 - Profile: Stats no longer care about the `Include Previously Watched` setting. All previously watched items will be counted in the stats now.
 - Activity: Added `index` to `WatchedID` column to speed up queries.
-- SeasonsListEpisode when `Hide Spoilers` is on:
+- SeasonsListEpisode when `Hide Spoilers` is on (thanks [@goestav]!):
   - Show episode spoilers if its status is `FINISHED`;
   - Allow changing status without showing spoilers (useful for when setting an episode to PLANNED, etc).
 
@@ -1791,3 +1791,4 @@ Welcome to Watcharr :popcorn:, hope it is enjoyed and improves anyone's experien
 [@Dredsen]: https://github.com/Dredsen
 [@GreatGatsby102]: https://github.com/GreatGatsby102
 [@4qu4r1um]: https://github.com/4qu4r1um
+[@goestav]: https://github.com/goestav